# Improve performance of Linux X509Chain building via caching