Yes, it introduces the most difficult of problems: cache invalidation.  Also, something got renamed.  And some second thing is probably worth mentioning.

X509Chain will use cached versions of CU\My, CU\CA, and CU\Root, in addition to the already cached LM\CA and LM\Root stores.

By avoiding the dips down to the filesystem it knocks a significant amount of I/O and computation off of the store reloads (~45% of one of my perf tests was spent doing SHA-1... for verifying the HMAC on opening the CU\My files).

The new heuristics are:
* CU\My, CU\CA, CU\Root
  * After one second elapses check to see if the LastWriteTimeUtc on the directory has changed, if so, invalidate the cache.
  * After 30 seconds invalidate the cache no matter what.
  * Cache invalidation is per store
* LM\CA, LM\Root
  * After 5 seconds elapses check to see if the LastWriteTimeUtc on either (or both) of the SSL_CERT_FILE file or SSL_CERT_DIR directory has changed (and if so, invalidate the cache)
  * After 5 minutes invalidate the cache no matter what.
  * System stores are rebuilt together.

All X509Store objects for LM\CA and LM\Root will get the same Pal instance (which just no-ops the Dispose) which is a projection over the cache.

All X509Store objects for CU\ stores will still use the existing "raw read" model, because otherwise the cache led to observable differences between Windows and Linux.  Only X509Chain gets the cached stores.

In running a chain-build-only 1000 times in a loop test, before this change my machine reported ~13.6ms per chain build, and after was 822ns.

In running an SslStream-to-SslStream handshake test, 10000 iterations before the change was 7 minutes 51 seconds.  After was 3 minutes 40 seconds.

This change also fixes a bug introduced in the previous perf enhancement where ERR_get_error was used instead of ERR_peek_last_error (wrong end of the queue, and destructive vs passive)... and that the managed code wasn't checking for the error anyways.

Fixes #35086.
Fixes #32576.