@sebastienros 2019-05-20
There has been some progress in the https handshake scenario that I am tracking, but they should not be related to these changes because the aspnet repos hasn't picked up a runtime with your changes yet. I will try to figure out which commits were responsible.

However I tried a runtime that included your improvements on Friday and I didn't see any change. I will work on it some more to be certain of it. I will also look at the traces to see what's the bottleneck.