@stephentoub 2019-05-20
> However I tried a runtime that included your improvements on Friday and I didn't see any change.

That is very surprising to me, assuming the "https handshake scenario" is primarily focused on the cost of setting up an SslStream connection with AuthenticateAsClient/ServerAsync.  I would expect this change to more than move the needle. In local tests, we saw this more than double the speed of that.

When you say runtime, I assume that includes both System.Security.Cryptography.X509Certificates.dll and System.Security.Cryptography.Native.so?  Those would both be necessary.