@sebastienros 2019-05-20
I even checked the code changes to be sure about what it's doing. That's why I said I would try again because it's very surprising. Ideally, instead of me picking up a dotnet runtime manually and hoping the SDK will respect it, I'd really prefer to have my hand on local dlls/so with the changes to be certain they are correct. I had some issues with the sdk last week.