@sebastienros 2019-05-20
The performance improved, and the charts are actually reflecting this change. I was confused because the version I was asking for and the ones actually used by the SDK were not consistent.

![image](https://user-images.githubusercontent.com/1165805/58058727-dbe58c00-7b1e-11e9-93aa-fc2e6c2a3a89.png)

Overall it's 2-3 times faster. However it's still 3 times slower than what is achieved on Windows or other plaftorms (for instance NodeJs).